Added reformatting for headings (WIP)
Headings inside patch notes will be replaced with <h3> or <h4> (major and minor headings). This is still heavily WIP as the patch notes do not follow a uniform html format. There are also some special cases which are improperly formatted wich will require a lot of cleanup.

A lot of problems have been annotated with todos, however the whole system will probably require a different approach as the heading format is pretty much randomized (in regard to the order of the html tags <em>, <strong> and <span>).