Improved script and small fixes.
Fixed a bug where load_new would delete the old cache.
Added a --copy_to parameter that can specify an additional path to copy the output file to.
Added 2 in 1 modes to load and create the html in one go. That way only one command line command is required to update everything.